Refactor into modules and automate sign-in
