+++ index.html
@@ -4,78 +4,8 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Personal YouTube Viewer</title>
-  <style>
-    body {
-      margin: 0;
-      font-family: Arial, sans-serif;
-      display: flex;
-      min-height: 100vh;
-    }
-    #sidebar {
-      background: #333;
-      color: #fff;
-      width: 200px;
-      transition: width 0.3s;
-      overflow: hidden;
-    }
-    #sidebar.collapsed {
-      width: 60px;
-    }
-    #sidebar button {
-      display: block;
-      width: 100%;
-      background: none;
-      border: none;
-      color: inherit;
-      padding: 12px;
-      text-align: left;
-      cursor: pointer;
-    }
-    #sidebar button:hover {
-      background: #444;
-    }
-    #toggleSidebar {
-      font-size: 24px;
-      text-align: center;
-      padding: 12px;
-      cursor: pointer;
-    }
-    #main {
-      flex: 1;
-      padding: 20px;
-    }
-    .view {
-      display: none;
-    }
-    .active {
-      display: block;
-    }
-    table {
-      border-collapse: collapse;
-      width: 100%;
-    }
-    th, td {
-      border: 1px solid #ddd;
-      padding: 8px;
-    }
-    th {
-      background-color: #f2f2f2;
-    }
-    #speedOverlay {
-      position: absolute;
-      top: 10px;
-      right: 10px;
-      background: rgba(0,0,0,0.7);
-      color: #fff;
-      padding: 5px 10px;
-      border-radius: 4px;
-      display: none;
-    }
-    #player {
-      width: 100%;
-      height: 70vh;
-    }
-  </style>
+  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
+  <link rel="stylesheet" href="styles/style.css">
 </head>
 <body>
   <nav id="sidebar">
@@ -86,9 +16,6 @@
     <button id="nav_search">Search</button>
   </nav>
   <div id="main">
-    <div id="signin">
-      <button id="authorize_button">Authorize</button>
-    </div>
     <div id="player_view" class="view">
       <div id="player"></div>
       <div id="speedOverlay"></div>
@@ -142,6 +69,6 @@
   </div>
   <script src="https://apis.google.com/js/api.js"></script>
   <script src="config.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="scripts/main.js"></script>
 </body>
 </html>
