+++ styles/style.css
@@ -1,0 +1,70 @@
+body {
+  margin: 0;
+  font-family: Arial, sans-serif;
+  display: flex;
+  min-height: 100vh;
+}
+#sidebar {
+  background: #333;
+  color: #fff;
+  width: 200px;
+  transition: width 0.3s;
+  overflow: hidden;
+}
+#sidebar.collapsed {
+  width: 60px;
+}
+#sidebar button {
+  display: block;
+  width: 100%;
+  background: none;
+  border: none;
+  color: inherit;
+  padding: 12px;
+  text-align: left;
+  cursor: pointer;
+}
+#sidebar button:hover {
+  background: #444;
+}
+#toggleSidebar {
+  font-size: 24px;
+  text-align: center;
+  padding: 12px;
+  cursor: pointer;
+}
+#main {
+  flex: 1;
+  padding: 20px;
+}
+.view {
+  display: none;
+}
+.active {
+  display: block;
+}
+table {
+  border-collapse: collapse;
+  width: 100%;
+}
+th, td {
+  border: 1px solid #ddd;
+  padding: 8px;
+}
+th {
+  background-color: #f2f2f2;
+}
+#speedOverlay {
+  position: absolute;
+  top: 10px;
+  right: 10px;
+  background: rgba(0,0,0,0.7);
+  color: #fff;
+  padding: 5px 10px;
+  border-radius: 4px;
+  display: none;
+}
+#player {
+  width: 100%;
+  height: 70vh;
+}
